requirements.txt
aiogram==2.25.2
aiohttp==3.8.6
aiosignal==1.4.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.10.5
charset-normalizer==3.4.3
frozenlist==1.8.0
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
mysql-connector-python==8.1.0
propcache==0.4.0
protobuf==4.21.12
python-dotenv==1.1.1
pytz==2025.2
typing_extensions==4.15.0
yarl==1.21.0
